Fixes as per comments

+++ src/main/java/core/basesyntax/Circle.java
@@ -29,7 +29,7 @@ public class Circle extends Figure {
         return radius;
     }
 
-    public void setRadius(int radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
 }
